Move terminating jobs logic from services to background task
